Throw for attempt to load navigation property on un-tracked entity
Issue #7388

Throwing for Query method as well because it would lead to the same issues as using Load directly--that is, fixup would not happen. If we implement explicit loading for un-tracked entities in the future, then it is conceivable that we would make Query work using the same mechanism.

Also checked other places where operations might behave differently based on entity state and added some more tests for Reload.